BaiTapKafka/Tutorial13/order-book-simulation/requirements.txt
aiokafka==0.7.1
altair==4.1.0
appnope==0.1.2
argon2-cffi==20.1.0
astor==0.8.1
async-generator==1.10
attrs==21.2.0
autopep8==1.5.7
backcall==0.2.0
base58==2.1.0
bleach==3.3.0
blinker==1.4
cachetools==4.2.2
certifi==2021.5.30
cffi==1.14.5
chardet==4.0.0
charset-normalizer==3.3.2
click==7.1.2
confluent-kafka==2.4.0
decorator==5.0.9
defusedxml==0.7.1
entrypoints==0.3
gitdb==4.0.7
GitPython==3.1.17
idna==2.10
ipykernel==5.5.5
ipython==7.24.1
ipython-genutils==0.2.0
ipywidgets==7.6.3
jedi==0.18.0
Jinja2==3.0.1
jsonschema==3.2.0
jupyter-client==6.1.12
jupyter-core==4.7.1
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.0
kafka-python @ git+https://github.com/dpkp/kafka-python.git@a6d0579d3cadd3826dd364b01bc12a2173139abc
markdown-it-py==3.0.0
MarkupSafe==2.0.1
matplotlib-inline==0.1.2
mdurl==0.1.2
mistune==0.8.4
nbclient==0.5.3
nbconvert==6.0.7
nbformat==5.1.3
nest-asyncio==1.5.1
notebook==6.4.0
numpy==1.26.4
packaging==20.9
pandas==2.2.2
pandocfilters==1.4.3
parso==0.8.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.2.0
prometheus-client==0.11.0
prompt-toolkit==3.0.18
protobuf==4.25.3
ptyprocess==0.7.0
pyarrow==16.1.0
pycodestyle==2.7.0
pycparser==2.20
pydeck==0.9.1
Pygments==2.18.0
pyparsing==2.4.7
pyrsistent==0.17.3
python-dateutil==2.9.0.post0
pytz==2021.1
pyzmq==26.0.3
requests==2.31.0
rich==13.7.1
Send2Trash==1.5.0
setuptools==69.5.1
six==1.16.0
smmap==4.0.0
streamlit==1.34.0
tenacity==8.3.0
terminado==0.10.0
testpath==0.5.0
toml==0.10.2
toolz==0.11.1
tornado==6.1
traitlets==5.0.5
typing_extensions==4.11.0
tzdata==2024.1
tzlocal==2.1
urllib3==1.26.5
validators==0.18.2
watchdog==2.1.2
wcwidth==0.2.5
webencodings==0.5.1
widgetsnbextension==3.5.1
